Use Alert repo to check for new ones
